Add validation styling and password check
Added styling for valid and invalid inputs, created a script that checks
if the typed in passwords are the same, if they are a green border is set
to both inputs, if not a red border is applied and a text appears below
"* Passwords do not match".
Linked "Create Account" button to the form via the id.
Added placeholder texts to show examples of correct data input.